backend/requirements.txt
# Core Framework
flask==2.3.3
sqlalchemy==2.0.20
apscheduler==3.10.4
flask-limiter==3.5.0
pydantic==2.3.0

# Web Scraping
playwright==1.38.0
playwright-stealth
beautifulsoup4==4.12.2
newspaper3k==0.2.8
trafilatura==1.6.1
pyyaml==6.0.1

# NLP and Vector Storage
sentence-transformers==2.2.2
huggingface-hub==0.16.4  # Specific version for compatibility
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0

# Testing
pytest==7.4.0
pytest-mock==3.11.1
coverage==7.3.1

# Frontend (Optional)
flask-cors==4.0.0